NIFI-8717 Refactoring PutHDFS
